Added API Key based authentication.
